FE-05: change component HotelPopulerList
