+++ src/components/HotelCard/HotelCard.scss
@@ -58,3 +58,14 @@
     }
   }
 }
+
+@media screen and (max-width: 334px) {
+  .room_card {
+    width: 15rem;
+    .room_card_img_wrapper {
+      .room_card_img {
+        height: 20rem;
+      }
+    }
+  }
+}
