+++ src/components/HotelPopulerList/HotelPopulerList.jsx
@@ -1,0 +1,46 @@
+import { Col, Row, Typography } from 'antd';
+import React from 'react';
+import { useEffect, useState } from 'react';
+
+import { HotelCard } from '..';
+import { hotelApi } from '../../api';
+import './HotelPopulerList.scss';
+
+const { Title } = Typography;
+
+const HotelPopulerList = () => {
+  const [data, setData] = useState([]);
+  const fetchHotels = async () => {
+    let result = await hotelApi.getAll();
+    setData(result);
+  };
+  useEffect(() => {
+    fetchHotels();
+  }, []);
+  return (
+    <div className="hotel_populer_list">
+      <Title className="title_section" level={2}>
+        Our Most Populer Room
+      </Title>
+      <Row className="content_section">
+        <Col span={12} offset={6}>
+          <p className="content_section_center">
+            It is a long established fact that a reader will be distracted by
+            the readable content of a page.
+          </p>
+        </Col>
+      </Row>
+      <Row className="hotel-list" style={{ justifyItems: 'center' }}>
+        {data.map((item, index) => {
+          return (
+            <Col xxl={8} xl={8} lg={8} md={12} sm={12} xs={24} key={index}>
+              <HotelCard {...item} />
+            </Col>
+          );
+        })}
+      </Row>
+    </div>
+  );
+};
+
+export default HotelPopulerList;
