+++ src/components/HotelPopulerList/HotelPopulerList.scss
@@ -1,0 +1,9 @@
+.hotel_populer_list {
+  & > h2 {
+    text-align: center;
+  }
+  .content_section {
+    text-align: center;
+    margin-bottom: 30px;
+  }
+}
