+++ src/components/index.js
@@ -1,5 +1,7 @@
 import HotelCard from './HotelCard/HotelCard';
+import HotelPopulerList from './HotelPopulerList/HotelPopulerList';
 import LargeButton from './LargeButton/LargeButton';
 
 export { LargeButton };
 export { HotelCard };
+export { HotelPopulerList };
