+++ src/pages/HomePage/HomePage.scss
@@ -1,12 +1,2 @@
 .homepage__container {
-  // background-color: red;
-  // margin: 60px 0;
-  // img {
-  //   width: 100%;
-  //   object-fit: contain;
-  // }
-  // justify-content: center;
-  & > h2 {
-    text-align: center;
-  }
 }
